feat: add Condomium adapter

+++ dapp/contracts/Condominium.sol
@@ -17,7 +17,7 @@ contract Condominium is ICondominium {
   constructor() {
     manager = msg.sender;
     for(uint8 i = 1; i<= 2; i++) { // blocks
-      for(uint16 j = 1; j <= 5; j++) { // floors
+      for(uint8 j = 1; j <= 5; j++) { // floors
         for(uint8 k = 1; k <= 5; k++) { // units
           unchecked {
             residences[(i * 1000) + (j * 100) + k] = true; // (1101) -> (2505)
